Script for extracting traffic values to vector
This is the main script that takes a bounding box for ROI and a road vector shapefile and returns a road vector with relative traffic conditions averaged over a given period of time. This works by mosaicking static maps from Bing Maps in ROI, classifying traffic conditions, averaging those conditions over the desired time period, and extracting the averaged traffic conditions to the underlying road vector.